+++ modules/etfmovies-videostream/src/main/java/com/etfmovies/videostream/service_interfaces/IVideoService.java
@@ -1,5 +1,5 @@
 package com.etfmovies.videostream.service_interfaces;
 
 public interface IVideoService {
-    String serveVideo(Long videoId);
+    String serveVideo(Integer videoId);
 }
